packages/cardpay-reward-scheduler/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Ian Calvert", email = "ian.calvert@cardstack.com"},
]
dependencies = [
    "boto3",
    "cloudpathlib[s3]>=0.7.1",
    "pyyaml>=6.0",
    "cachetools>=5.0.0",
    "pyarrow>=8.0.0",
    "web3>=5.29.1",
    "sentry-sdk>=1.5.12",
    "python-dotenv>=0.20.0",
    "schedule>=1.1.0",
    "duckdb>=0.3.4",
]
requires-python = ">=3.9"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "black>=22.3.0",
]
[tool.pdm]
[[tool.pdm.source]]
name = "pypi"
url = "https://pypi.org/simple"
verify_ssl = true


[tool.pdm.dev-dependencies]
[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"
